Completed DAOs

+++ GestioneSegreteriaLezione/src/main/java/it/unical/mat/webcomp21/persistence/dao/CorsoDiLaureaDAO.java
@@ -9,11 +9,11 @@ import it.unical.mat.webcomp21.model.Studente;
 
 public interface CorsoDiLaureaDAO {
 	public void save(CorsoDiLaurea cdl);  // Create
-	public CorsoDiLaurea findByPrimaryKey(int id);     // Retrieve
+	public CorsoDiLaurea findByPrimaryKey(Long id);     // Retrieve
 	public List<CorsoDiLaurea> findAll();       
 	public void update(CorsoDiLaurea cdl); //Update
 	public void delete(CorsoDiLaurea cdl); //Delete	
 	
 	
-	public ArrayList<Studente> findStudentiByCorsoDiLaureaId(int id);
+	public ArrayList<Studente> findStudentiByCorsoDiLaureaId(Long id);
 }
